Merged PR 267: Unilateral JSON conversion properties
Main idea: The macro which defines the implementation of a `jsonProperties` tuple is now public:

```diff
- #    define OSSIACO_INTERNAL_JSON_PROPERTIES_IMPL(_chainedCall)                                    \
+ #    define OSSIACO_JSON_PROPERTIES_IMPL(_chainedCall)
```

Users can now call this macro (and nothing more) to define a tuple of JSON properties. This is sufficient to be able to serialize a class with a call to `JsonSerializer`.

In addition to defining JSON properties with pointer to member functions, users can now specify a pointer to a getter.

This has been accomplished by a change to `PropertyConverter` (now called `JsonProperty`) with a similar rename for the new class `ChronoFmtJsonProperty`.

Rather than unconditionally providing both `toJson` and `fromJson`, SFINAE is used to expose these methods if and only if they make sense for the type of member pointer passed:
- a const-qualified getter or a pointer to member data enables `toJson`
- a pointer to member data enables `fromJson`

A future PR will add a similar functionality for setter methods.

Note the version of Catch2 has been bumped to 2.4.2 to make use of the `STATIC_REQUIRE` macro also