Merged PR 282: Use centralized repo of azure templates
Uses CI templates in this repo as jumping-off point for creation of https://dev.azure.com/ossiaco/ossiaco/_git/AzureTemplates

In-scope for this PR is the use of these templates and the templates themselves; for issues with the templates themselves please record them as a comment next to a line in this codebase which uses or illustrates the issue you have with the template code.

Note that we currently just grab `refs/heads/master` by [default](https://docs.microsoft.com/en-us/azure/devops/pipelines/yaml-schema?view=vsts&tabs=schema#repository). Once this PR passes the current contents of this repo will be tagged as a release called `0.0.1-alpha-converter` or some such. The idea is that we will integrate the templates into one repo at a time, making changes as necessary.

N.B.: The change to `vocab_type.hpp` is a happy accident in which adding a test for XCode debug build allowed me to figure out a replacement for some brittle, platform-specific workaround.