Merged PR 306: Compile-time efficiency measures
Uses `hana::basic_tuple` as the main class for building properties, and adds a means to adapt classes that evaluates to a single call to `hana::make_basic_tuple`.

See the second chart here http://boostorg.github.io/hana/index.html#tutorial-performance-compile
From the [Hana documentation](http://boostorg.github.io/hana/structboost_1_1hana_1_1basic__tuple.html),
> Whereas hana::tuple aims to provide an interface somewhat close to a std::tuple, basic_tuple provides the strict minimum required to implement a closure with maximum compile-time efficiency.

This reduced interface is adequate for the task of defining and building collections of properties.

Happily, it allows us to remove the macro-based workarounds for issues with MSVC too.

Pursuant to the goal of improving compile-time efficiency, we also add an option to define JSON properties in one fell swoop rather than with evaluation to repeated calls to `append`. The unit tests include some simple checks of this new functionality.

In the process of writing the unit tests I found an oversight in the implementation of `printTypeName` which has been fixed.