Merged PR 318: Add poly registration macro
Add a convenience macro for defining a variable which stores the result of the converter factory registration method.

This macro is intended for library users and does not appear in the examples or unit tests, as it would hurt readability and/or make it harder to `assert` or (Catch) `REQUIRE` that registration has taken place.